+++ Sources/SwiftSDK/Geo/BackendlessGeoQuery.swift
@@ -27,26 +27,96 @@
     case FEET = 4
 }
 
-@objcMembers open class BackendlessGeoQuery: NSObject {
+@objcMembers open class BackendlessGeoQuery: NSObject, NSCoding, Codable {
     
     open var geoPoint: GeoPoint?
-    open var radius: NSNumber?
+    open var radius: Double?
     open var categories: [String]?
     open var includemetadata = false
-    open var metadata: [String : Any]?
+    
+    private var _metadata: JSON?
+    open var metadata: [String : Any]? {
+        get {
+            return _metadata?.dictionaryObject
+        }
+        set(newMetadata) {
+            if let newMetadata = newMetadata {
+                _metadata = JSON(newMetadata)
+            }
+        }
+    }
+    
     open var whereClause: String?
     open var rectangle: GeoQueryRectangle?
     open var pageSize: Int = 10
     open var offset: Int = 0
-    open var clustergridsize: NSNumber?
     
     open private(set) var degreePerPixel: Double = 0.0
-    open private(set) var clusterGridSize: Int = 100
+    open private(set) var clusterGridSize: Double = 100.0
     
-    private var units: NSNumber?
+    private var units: Int?
+    
+    enum CodingKeys: String, CodingKey {
+        case geoPoint
+        case radius
+        case categories
+        case includemetadata
+        case _metadata = "metadata"
+        case whereClause
+        case rectangle
+        case pageSize
+        case offset
+        case degreePerPixel
+        case clusterGridSize
+    }
+    
+    public override init() { }
+    
+    public init(geoPoint: GeoPoint?, radius: Double?, categories: [String]?, includemetadata: Bool, _metadata: JSON?, whereClause: String?, rectangle: GeoQueryRectangle?, pageSize: Int, offset: Int, degreePerPixel: Double, clusterGridSize: Double) {
+        self.geoPoint = geoPoint
+        self.radius = radius
+        self.categories = categories
+        self.includemetadata = includemetadata
+        self._metadata = _metadata
+        self.whereClause = whereClause
+        self.rectangle = rectangle
+        self.pageSize = pageSize
+        self.offset = offset
+        self.degreePerPixel = degreePerPixel
+        self.clusterGridSize = clusterGridSize
+    }
+    
+    convenience public required init?(coder aDecoder: NSCoder) {
+        let geoPoint = aDecoder.decodeObject(forKey: CodingKeys.geoPoint.rawValue) as? GeoPoint
+        let radius = aDecoder.decodeDouble(forKey: CodingKeys.radius.rawValue)
+        let categories = aDecoder.decodeObject(forKey: CodingKeys.categories.rawValue) as? [String]
+        let includemetadata = aDecoder.decodeBool(forKey: CodingKeys.includemetadata.rawValue)
+        let _metadata = aDecoder.decodeObject(forKey: CodingKeys._metadata.rawValue) as? JSON
+        let whereClause = aDecoder.decodeObject(forKey: CodingKeys.whereClause.rawValue) as? String
+        let rectangle = aDecoder.decodeObject(forKey: CodingKeys.rectangle.rawValue) as? GeoQueryRectangle
+        let pageSize = aDecoder.decodeInteger(forKey: CodingKeys.pageSize.rawValue)
+        let offset = aDecoder.decodeInteger(forKey: CodingKeys.offset.rawValue)
+        let degreePerPixel = aDecoder.decodeDouble(forKey: CodingKeys.degreePerPixel.rawValue)
+        let clusterGridSize = aDecoder.decodeDouble(forKey: CodingKeys.clusterGridSize.rawValue)
+        self.init(geoPoint: geoPoint, radius: radius, categories: categories, includemetadata: includemetadata, _metadata: _metadata, whereClause: whereClause, rectangle: rectangle, pageSize: pageSize, offset: offset, degreePerPixel: degreePerPixel, clusterGridSize: clusterGridSize)
+    }
+    
+    public func encode(with aCoder: NSCoder) {
+        aCoder.encode(geoPoint, forKey: CodingKeys.geoPoint.rawValue)
+        aCoder.encode(radius, forKey: CodingKeys.radius.rawValue)
+        aCoder.encode(categories, forKey: CodingKeys.categories.rawValue)
+        aCoder.encode(includemetadata, forKey: CodingKeys.includemetadata.rawValue)
+        aCoder.encode(_metadata, forKey: CodingKeys._metadata.rawValue)
+        aCoder.encode(whereClause, forKey: CodingKeys.whereClause.rawValue)
+        aCoder.encode(rectangle, forKey: CodingKeys.rectangle.rawValue)
+        aCoder.encode(pageSize, forKey: CodingKeys.pageSize.rawValue)
+        aCoder.encode(offset, forKey: CodingKeys.offset.rawValue)
+        aCoder.encode(degreePerPixel, forKey: CodingKeys.degreePerPixel.rawValue)
+        aCoder.encode(clusterGridSize, forKey: CodingKeys.clusterGridSize.rawValue)
+    }
     
     open func setUnits(units: Int) {
-        self.units = NSNumber(integerLiteral: units)
+        self.units = units
     }
     
     open func getUnits() -> String {
@@ -68,7 +138,7 @@
         return "MILES"
     }
     
-    open func setClusteringParams(degreePerPixel: Double, clusterGridSize: Int) {
+    open func setClusteringParams(degreePerPixel: Double, clusterGridSize: Double) {
         self.degreePerPixel = degreePerPixel
         self.clusterGridSize = clusterGridSize
     }
@@ -77,7 +147,7 @@
         setClusteringParams(westLongitude: westLongitude, eastLongitude: eastLongitude, mapWidth: mapWidth, clusterGridSize: 100)
     }
     
-    open func setClusteringParams(westLongitude: Double, eastLongitude: Double, mapWidth: Int, clusterGridSize: Int) {
+    open func setClusteringParams(westLongitude: Double, eastLongitude: Double, mapWidth: Int, clusterGridSize: Double) {
         if eastLongitude - westLongitude < 0 {
             self.degreePerPixel = ((eastLongitude - westLongitude) + 360) / Double(mapWidth)
         }