+++ Sources/SwiftSDK/Messaging/RT/RTMessaging.swift
@@ -134,10 +134,12 @@ class RTMessaging: RTListener {
     
     func addMessageListener(selector: String?, responseHandler: ((PublishMessageInfo) -> Void)!, errorHandler: ((Fault) -> Void)!) -> RTSubscription? {
         let wrappedBlock: (Any) -> () = { response in
+            
             if let response = response as? [String : Any] {
-                let publishMessageInfo = self.processResponse.adaptToPublishMessageInfo(messageInfoDictionary: response)
-                responseHandler(publishMessageInfo)
+                let publishMessageInfo = self.processResponse.adapt(responseDictionary: response, to: PublishMessageInfo.self)
+                responseHandler(publishMessageInfo as! PublishMessageInfo)
             }
+            
         }
         if self.channel.isJoined {
             var options = [String : Any]()
@@ -150,9 +152,7 @@ class RTMessaging: RTListener {
             subscription.subscribe()
             return subscription
         }
-        else {
-            return addWaitingSubscription(event: PUB_SUB_MESSAGES, selector: selector, responseHandler: wrappedBlock, errorHandler: errorHandler)
-        }
+        return addWaitingSubscription(event: PUB_SUB_MESSAGES, selector: selector, responseHandler: wrappedBlock, errorHandler: errorHandler)
     }
     
     func removeMessageListeners(selector: String?) {